services/api-service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-dotenv
passlib[bcrypt]
pydantic[email]

